+++ tests/specs/api/stripe.spec.js
@@ -2,10 +2,8 @@ import * as httpMock from '../../mocks/http';
 import { getAuthenticationHeaders } from '../../../src/utils/auth';
 import { StripeApi } from '../../../src/api/stripe';
 
-
 describe('Stripe API', () => {
   const serviceUrl = 'http://some-url.com';
-  const userId = 'user-id';
   const httpHeaders = getAuthenticationHeaders({
     jwt: 'jwt'
   });
@@ -21,69 +19,28 @@ describe('Stripe API', () => {
     httpMock.restore();
   });
 
-  describe('#updateCustomer', () => {
+  describe('#getAccount', () => {
+
     it('should call http', () => {
-      return api.updateCustomer(userId, 'token').then(() => {
-        const fullUrl = api.getFullURL('appUsers', userId, 'stripe', 'customer');
-        httpSpy.should.have.been.calledWith('POST', fullUrl, {
-          token: 'token'
-        }, httpHeaders);
+      console.log(api.getAccount);
+      return api.getAccount().then(() => {
+        const fullUrl = api.getFullURL('stripe', 'account');
+        httpSpy.should.have.been.calledWith('GET', fullUrl, {}, httpHeaders);
       });
     });
 
-    it('should throw if no token provided', () => {
-      return api.updateCustomer(userId).catch(() => {
-        httpSpy.should.not.have.been.called;
+    it('should accept appToken as authentication method', () => {
+      const httpHeaders = getAuthenticationHeaders({
+        appToken: 'appToken'
       });
-    });
-    
-    describe('with app-token', () => {
-      it('should throw', () => {
-        let badApi = new StripeApi(serviceUrl, getAuthenticationHeaders({
-          appToken: 'token'
-        }));
+      api = new StripeApi(serviceUrl, httpHeaders);
 
-        return badApi.updateCustomer(userId, 'token').catch(() => {
-          httpSpy.should.not.have.been.called;
+      return api.getAccount()
+        .then(() => {
+          const fullUrl = api.getFullURL('stripe', 'account');
+          httpSpy.should.have.been.calledWith('GET', fullUrl, {}, httpHeaders);
         });
-      })
-    });
-  });
-
-  describe('#createTransaction', () => {
-    describe('with token', () => {
-      it('should call http', () => {
-        return api.createTransaction(userId, 'actionId', 'token').then(() => {
-          const fullUrl = api.getFullURL('appUsers', userId, 'stripe', 'transaction');
-          httpSpy.should.have.been.calledWith('POST', fullUrl, {
-            actionId: 'actionId',
-            token: 'token'
-          }, httpHeaders);
-        });
-      });
-
-      it('should throw if no actionId provided', () => {
-        return api.createTransaction(userId, undefined, 'token').catch(() => {
-          httpSpy.should.not.have.been.called;
-        });
-      });
     });
 
-    describe('without token', () => {
-      it('should call http', () => {
-        return api.createTransaction(userId, 'actionId').then(() => {
-          const fullUrl = api.getFullURL('appUsers', userId, 'stripe', 'transaction');
-          httpSpy.should.have.been.calledWith('POST', fullUrl, {
-            actionId: 'actionId'
-          }, httpHeaders);
-        });
-      });
-
-      it('should throw if no actionId provided', () => {
-        return api.createTransaction(userId, undefined).catch(() => {
-          httpSpy.should.not.have.been.called;
-        });
-      });
-    });
   });
 });
